Editor: extract caption buttons widget and route actions through controllers
Extract titlebar caption controls into a dedicated EditorCaptionButtons component.

Keep EditorTitleBar focused on composition/forwarding, while EditorNode and ProjectManager handle minimize/maximize/close behavior.

This preserves Windows unified-titlebar behavior while improving separation of UI and platform/controller responsibilities.
